+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - added support for `externalIds` in `/packages.search.id` ([#26][gui26])
   - output format changes in `/packages.search` and `/packages.search.id`
   - message `/prompt/choice/update/variant` now informs about `previouslySelectedValue` and `importedValues` if applicable
+  - added new `/plugins.export` endpoint
 
 [gui26]: https://github.com/memo33/sc4pac-gui/issues/26